Add a random token generator fn
The existing code manually builds up tokens in a couple places, with some casts and indexing in the process.

This change extracts token generation from an alphabet to a helper function, updates the types to avoid casting, and refactors it to use `choose_multiple` from rand0.8's `SliceRandom` to avoid direct indexing.

Ultimately it's only a bit more than a cosmetic change, so feel free to close/reject if you prefer keeping a flatter style with fewer helper functions.

---

Note that `rand 0.9` has been out since January and it might be worth updating. From uses I've spotted, AIP is only affected by names changing in 0.9 (like `rand::thread_rng()` becomes just `rand::rng()`), so it's pretty mechanical to update. I'd be happy to follow up with that if you want!